fix(folderlisting): fix the layout so file entry spans the width of the panel
The folder listing to the right of the sidebar was centering the file list. There was additional
styling from the original frame of the page that was interacting poorly with the material UI drawer
setup. I have moved that structure down inside the main panel so it surrounds the editor widget,
keeping the layout as is for the editor.