feat(sidebar): populate username from textile peer
The store now pulls the profile info from the textile peer on first boot. That is the name that is
displayed, defaulting to Anon if none has been set. The profile picture is not being displayed for
now. I need to figure out how textile expects image files to be accessed and displayed.