fix(posts): Show post on screen after clicking `Post` button
There were several problems going on here. The mobx store wasn't tracking posts. Thoses posts weren't being observed in the Group.tsx component, and the magical hilarity that is mobx observable binding had to be figured out.

These are all excuses for a commit that is doing too much, but a few key points. We now have a currentPosts collection in the store, this is updated on add posts. That collection is observed in the Group.tsx component.

I also found a weirdness with react hooks inside observed function components. I have managed to dodge the issue by putting effects and observations at different levels in the heirarchy.